refactor(metatab): isolate form
